Works for 3 AddOrder Messages.
Some code cleanup

Small change - vi-scoped fifo is now sending the order cluster to the test-harness, now test-harness has to deserialize it.

Another check point.

Going to refactor the OrderBook-Top-Level.vi now to accept a struct as input instead of the deserialized version of the command.
Will only support AddOrder and GetAllOrders for this iteration.

Checkpoint - synch